Instructions and Need to Know Information
Please read this file if you actually want to know what is going on. Questions such as opening files and how to run the program are explained in this write-up. Since we have a wide range of experience in the class it crucial that those who are new to programming take some time to go through the resources